🧪 Sentinel: add useAssistant coverage
